requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.3
